# openapi.yaml
openapi: 3.1.0
info:
  title: EventOps AI
  description: |
    AI-штаб для координации команды на мероприятии.

    **Аутентификация:** все эндпоинты кроме `/auth/login` требуют
    заголовок `Authorization: Bearer <token>`.

    **Visibility-фильтрация:** бэкенд автоматически фильтрует тикеты
    и сообщения по правам текущего сотрудника. Фронт ничего
    дополнительно не фильтрует.
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Local dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Enums ────────────────────────────────────────────────────────────────
  schemas:
    TicketType:
      type: string
      enum: [planned, incident, tech, question]
      description: |
        - `planned`  — плановая задача
        - `incident` — инцидент
        - `tech`     — техническая проблема
        - `question` — вопрос
    TicketPriority:
      type: string
      enum: [low, medium, high, critical]
    TicketStatus:
      type: string
      enum: [new, in_progress, waiting, resolved, closed]
      description: |
        - `waiting` — ожидает подтверждения координатора
    Visibility:
      type: string
      enum: [public, role_only, confidential]
      description: |
        - `public`       — видят все участники мероприятия
        - `role_only`    — видит только своя роль + admin
        - `confidential` — только admin + роли с can_see_confidential
    StaffStatus:
      type: string
      enum: [free, busy, on_task, offline]
...
    TicketTarget:
      type: object
      description: Нормализованный target задачи в ответе
      required: [all, role_ids, staff_ids]
      properties:
        all:
          type: boolean
          default: false
        role_ids:
          type: array
          items:
            type: integer
          default: []
        staff_ids:
          type: array
          items:
            type: integer
          default: []
    TicketTargetPatch:
      type: object
      nullable: true
      description: |
        Входной target для create/update.
        Бэкенд нормализует значение в формат `TicketTarget`.
      properties:
        all:
          type: boolean
        role_ids:
          type: array
          items:
            type: integer
        staff_ids:
          type: array
          items:
            type: integer
      additionalProperties: true

    # ─── Zone ───────────────────────────────────────────────────────────────
...
    StaffShort:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          $ref: '#/components/schemas/StaffStatus'
...
    TicketCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: На регистрации очередь
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/TicketType'
          default: incident
        priority:
          $ref: '#/components/schemas/TicketPriority'
          default: medium
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: public
        assignee_role_id:
          type: integer
          nullable: true
        target:
          $ref: '#/components/schemas/TicketTargetPatch'
          description: |
            Целевая аудитория задачи (все / роли / конкретные сотрудники).
            При отсутствии бэкенд формирует target автоматически.
        previous_messages:
          type: array
          nullable: true
          description: |
            Только для ручного создания/импорта тикета. В agent flow фронт не заполняет это поле:
            бэкенд сам ведёт отдельную цепочку для каждого authenticated staff внутри мероприятия.
            Каждый элемент может содержать `role`, `text`, `source`, `audio_file`.
          items:
            type: object
            additionalProperties: true
    TicketUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/TicketType'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        visibility:
          $ref: '#/components/schemas/Visibility'
        assignee_role_id:
          type: integer
          nullable: true
        target:
          $ref: '#/components/schemas/TicketTargetPatch'
          description: Обновление целевой аудитории задачи
        previous_messages:
          type: array
          nullable: true
          description: Обновление цепочки исходных сообщений задачи
          items:
            type: object
            additionalProperties: true
    TicketAssignmentOut:
      type: object
      required: [id, staff, assigned_at, confirmed]
      properties:
        id:
          type: integer
        staff:
          $ref: '#/components/schemas/StaffShort'
        assigned_at:
          type: string
          format: date-time
        confirmed:
          type: boolean
    TicketReplyCreate:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          example: Взял задачу, буду через 5 минут
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: public
    TicketReply:
      type: object
      required: [id, event_id, ticket_id, from_staff_id, content, visibility, created_at]
      properties:
        id:
          type: integer
        event_id:
          type: integer
        ticket_id:
          type: integer
        from_staff_id:
          type: integer
        sender:
          $ref: '#/components/schemas/StaffShort'
          nullable: true
        content:
          type: string
        visibility:
          $ref: '#/components/schemas/Visibility'
        created_at:
          type: string
          format: date-time
    DocumentAttachment:
      type: object
      required: [id, title, filename, size, path, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        filename:
          type: string
        content_type:
          type: string
          nullable: true
        size:
          type: integer
        path:
          type: string
          description: Локальный путь к сохранённому файлу
        uploaded_by_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
    Ticket:
      type: object
      required: [id, event_id, title, type, priority, status, visibility, created_at, updated_at]
      properties:
        id:
          type: integer
        event_id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/TicketType'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        visibility:
          $ref: '#/components/schemas/Visibility'
        created_by_id:
          type: integer
          nullable: true
        sender:
          $ref: '#/components/schemas/StaffShort'
          nullable: true
          description: Отправитель/создатель тикета
        created_by:
          $ref: '#/components/schemas/StaffShort'
          nullable: true
        assignee_role_id:
          type: integer
          nullable: true
        target:
          $ref: '#/components/schemas/TicketTarget'
          description: Нормализованный target задачи
          default:
            all: false
            role_ids: []
            staff_ids: []
        previous_messages:
          type: array
          description: |
            Все сообщения, которые привели к созданию задачи: тексты, источник и путь к аудиофайлу,
            если команда пришла голосом. В agent flow цепочка ведётся бэкендом отдельно для каждого
            пользователя (`event_id + current_staff`) и сбрасывается после создания задачи или закрытия вопроса.
          items:
            type: object
            additionalProperties: true
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentAttachment'
        ai_suggestion:
          type: object
          nullable: true
          description: |
            Внутренний JSON с рекомендацией Алисы.
            Может содержать `reasoning`, `suggested_staff_ids`, `confidence`, `target` и др.
          additionalProperties: true
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/TicketAssignmentOut'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AssignRequest:
      type: object
      required: [staff_ids]
      properties:
        staff_ids:
          type: array
          items:
            type: integer
          minItems: 1
          example: [3, 7]
    ConfirmAssignmentRequest:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean

    # ─── Message ────────────────────────────────────────────────────────────
...
    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: Not found

# ─── Paths ───────────────────────────────────────────────────────────────────
paths:
...
  /events/{event_id}/tickets:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [tickets]
      summary: Список тикетов
      description: Фильтруется по visibility текущего пользователя
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/TicketPriority'
        - name: role_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TicketType'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
    post:
      tags: [tickets]
      summary: Создать тикет
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /events/{event_id}/tickets/{ticket_id}:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [tickets]
      summary: Тикет по id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '403':
          description: Нет доступа (visibility)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
    patch:
      tags: [tickets]
      summary: Обновить тикет
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /events/{event_id}/tickets/{ticket_id}/assign:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags: [tickets]
      summary: Назначить людей на тикет
      description: Ручное назначение. После подтверждения Алисы используй /agent/confirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /events/{event_id}/tickets/{ticket_id}/replies:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [tickets]
      summary: Ответы на тикет
      description: Доступно только пользователям, которые видят сам тикет
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketReply'
    post:
      tags: [tickets]
      summary: Добавить ответ к тикету
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketReplyCreate'
      responses:
        '201':
          description: Создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketReply'
  /events/{event_id}/tickets/{ticket_id}/documents:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags: [tickets]
      summary: Прикрепить документ к тикету
      description: Multipart upload. Доступно пользователям, которые видят тикет.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Документ прикреплён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentAttachment'
  /events/{event_id}/tickets/{ticket_id}/assignments/{assignment_id}:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      - name: ticket_id
        in: path
        required: true
        schema:
          type: integer
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [tickets]
      summary: Подтвердить или отклонить назначение
      description: Волонтёр подтверждает что принял задачу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmAssignmentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketAssignmentOut'

  # ── Knowledge Base ───────────────────────────────────────────────────────
...